implicit token in operations : working but making fail other unit tests
